+++ routes/api.php
@@ -4,4 +4,6 @@ Route::get('/notes', 'NotesController@getNotes');
 Route::post('/notes', 'NotesController@addNote');
 Route::delete('/notes', 'NotesController@deleteNote');
 Route::patch('/notes/{note}/pin', 'NotesController@pinNote');
+Route::patch('/notes/{note}/complete', 'NotesController@completeNote');
 Route::patch('/notes/{note}', 'NotesController@editNote');
+Route::get('/users', 'NotesController@getAssignableUsers');